Add files via upload
This is the 1st assignment and questions are given below.
Q1. Write a Python program to get the Fibonacci series between 0 to 50
Q2. Write a Python program that accepts a word from the user and reverse it.
Q3. Write a Python program to count the number of even and odd numbers from a series of numbers.